Move mapping-tree-view component to separate folder
