+++ src/ggrc-client/js/components/mapping-tree-view/mapping-tree-view.mustache
